changed: put import/export videodatabase methods in separate file
