+++ xbmc/video/database/CMakeLists.txt
@@ -1,4 +1,5 @@
 set(SOURCES VideoDatabase.cpp
+            VideoDatabaseExportImport.cpp
             VideoDatabaseSettings.cpp)
 
 set(HEADERS VideoDatabase.h)
